Use policy register
