--- apps/web/src/app/admin-console/core/policy-list.service.ts
@@ -1,13 +1,0 @@
-import { BasePolicy } from "../organizations/policies";
-
-export class PolicyListService {
-  private policies: BasePolicy[] = [];
-
-  addPolicies(policies: BasePolicy[]) {
-    this.policies.push(...policies);
-  }
-
-  getPolicies(): BasePolicy[] {
-    return this.policies;
-  }
-}
