+++ apps/web/src/app/admin-console/organizations/policies/index.ts
@@ -1,14 +1,2 @@
-export { BasePolicy, BasePolicyComponent } from "./base-policy.component";
-export { DisableSendPolicy } from "./disable-send.component";
-export { MasterPasswordPolicy } from "./master-password.component";
-export { PasswordGeneratorPolicy } from "./password-generator.component";
-export { vNextOrganizationDataOwnershipPolicy } from "./vnext-organization-data-ownership.component";
-export { OrganizationDataOwnershipPolicy } from "./organization-data-ownership.component";
-export { RequireSsoPolicy } from "./require-sso.component";
-export { ResetPasswordPolicy } from "./reset-password.component";
-export { SendOptionsPolicy } from "./send-options.component";
-export { SingleOrgPolicy } from "./single-org.component";
-export { TwoFactorAuthenticationPolicy } from "./two-factor-authentication.component";
 export { PoliciesComponent } from "./policies.component";
-export { RemoveUnlockWithPinPolicy } from "./remove-unlock-with-pin.component";
-export { RestrictedItemTypesPolicy } from "./restricted-item-types.component";
+export { policyRegister, POLICY_REGISTER_TOKEN } from "./policy-register";
