+++ apps/web/src/app/admin-console/organizations/policies/policy-list.service.ts
@@ -1,0 +1,13 @@
+import { BasePolicy } from "./base-policy.component";
+
+export class PolicyListService {
+  private policies: readonly BasePolicy[];
+
+  constructor(policies: BasePolicy[]) {
+    this.policies = Object.freeze([...policies]);
+  }
+
+  getPolicies() {
+    return this.policies;
+  }
+}
