+++ apps/web/src/app/admin-console/organizations/policies/policy-register.ts
@@ -1,0 +1,38 @@
+import { SafeInjectionToken } from "@bitwarden/ui-common";
+
+import { BasePolicy } from "./base-policy.component";
+import { DisableSendPolicy } from "./disable-send.component";
+import { MasterPasswordPolicy } from "./master-password.component";
+import { OrganizationDataOwnershipPolicy } from "./organization-data-ownership.component";
+import { PasswordGeneratorPolicy } from "./password-generator.component";
+import { RemoveUnlockWithPinPolicy } from "./remove-unlock-with-pin.component";
+import { RequireSsoPolicy } from "./require-sso.component";
+import { ResetPasswordPolicy } from "./reset-password.component";
+import { RestrictedItemTypesPolicy } from "./restricted-item-types.component";
+import { SendOptionsPolicy } from "./send-options.component";
+import { SingleOrgPolicy } from "./single-org.component";
+import { TwoFactorAuthenticationPolicy } from "./two-factor-authentication.component";
+import { vNextOrganizationDataOwnershipPolicy } from "./vnext-organization-data-ownership.component";
+
+
+
+export const POLICY_REGISTER_TOKEN = new SafeInjectionToken<BasePolicy[]>("POLICY_REGISTER_TOKEN");
+
+/**
+ * The policy register for OSS policies.
+ * Add your policy definition here if it is under the OSS license.
+ */
+export const policyRegister: BasePolicy[] = [
+  new TwoFactorAuthenticationPolicy(),
+  new MasterPasswordPolicy(),
+  new RemoveUnlockWithPinPolicy(),
+  new ResetPasswordPolicy(),
+  new PasswordGeneratorPolicy(),
+  new SingleOrgPolicy(),
+  new RequireSsoPolicy(),
+  new OrganizationDataOwnershipPolicy(),
+  new vNextOrganizationDataOwnershipPolicy(),
+  new DisableSendPolicy(),
+  new SendOptionsPolicy(),
+  new RestrictedItemTypesPolicy(),
+];
