+++ apps/web/src/app/app.component.ts
@@ -30,22 +30,6 @@ import { SearchService } from "@bitwarden/common/vault/abstractions/search.servi
 import { DialogService, ToastService } from "@bitwarden/components";
 import { KeyService, BiometricStateService } from "@bitwarden/key-management";
 
-import { PolicyListService } from "./admin-console/core/policy-list.service";
-import {
-  DisableSendPolicy,
-  MasterPasswordPolicy,
-  PasswordGeneratorPolicy,
-  OrganizationDataOwnershipPolicy,
-  vNextOrganizationDataOwnershipPolicy,
-  RequireSsoPolicy,
-  ResetPasswordPolicy,
-  SendOptionsPolicy,
-  SingleOrgPolicy,
-  TwoFactorAuthenticationPolicy,
-  RemoveUnlockWithPinPolicy,
-  RestrictedItemTypesPolicy,
-} from "./admin-console/organizations/policies";
-
 const BroadcasterSubscriptionId = "AppComponent";
 const IdleTimeout = 60000 * 10; // 10 minutes
 
@@ -79,7 +63,6 @@ export class AppComponent implements OnDestroy, OnInit {
     private serverNotificationsService: ServerNotificationsService,
     private stateService: StateService,
     private eventUploadService: EventUploadService,
-    protected policyListService: PolicyListService,
     protected configService: ConfigService,
     private dialogService: DialogService,
     private biometricStateService: BiometricStateService,
@@ -238,21 +221,6 @@ export class AppComponent implements OnDestroy, OnInit {
         }
       });
     });
-
-    this.policyListService.addPolicies([
-      new TwoFactorAuthenticationPolicy(),
-      new MasterPasswordPolicy(),
-      new RemoveUnlockWithPinPolicy(),
-      new ResetPasswordPolicy(),
-      new PasswordGeneratorPolicy(),
-      new SingleOrgPolicy(),
-      new RequireSsoPolicy(),
-      new OrganizationDataOwnershipPolicy(),
-      new vNextOrganizationDataOwnershipPolicy(),
-      new DisableSendPolicy(),
-      new SendOptionsPolicy(),
-      new RestrictedItemTypesPolicy(),
-    ]);
   }
 
   ngOnDestroy() {
