+++ apps/web/src/app/core/core.module.ts
@@ -113,7 +113,7 @@ import { DefaultSshImportPromptService, SshImportPromptService } from "@bitwarde
 import { WebOrganizationInviteService } from "@bitwarden/web-vault/app/auth/core/services/organization-invite/web-organization-invite.service";
 
 import { flagEnabled } from "../../utils/flags";
-import { PolicyListService } from "../admin-console/core/policy-list.service";
+import { POLICY_REGISTER_TOKEN, policyRegister } from "../admin-console/organizations/policies";
 import {
   WebChangePasswordService,
   WebRegistrationFinishService,
@@ -152,7 +152,10 @@ const safeProviders: SafeProvider[] = [
   safeProvider(InitService),
   safeProvider(RouterService),
   safeProvider(EventService),
-  safeProvider(PolicyListService),
+  safeProvider({
+    provide: POLICY_REGISTER_TOKEN,
+    useValue: policyRegister,
+  }),
   safeProvider({
     provide: DEFAULT_VAULT_TIMEOUT,
     deps: [PlatformUtilsService],
