+++ bitwarden_license/bit-web/src/app/admin-console/policies/disable-personal-vault-export.component.ts
@@ -1,10 +1,12 @@
 import { Component } from "@angular/core";
 
+
 import { PolicyType } from "@bitwarden/common/admin-console/enums";
 import {
   BasePolicy,
   BasePolicyComponent,
 } from "@bitwarden/web-vault/app/admin-console/organizations/policies/base-policy.component";
+import { SharedModule } from "@bitwarden/web-vault/app/shared";
 
 export class DisablePersonalVaultExportPolicy extends BasePolicy {
   name = "disablePersonalVaultExport";
@@ -14,8 +16,7 @@ export class DisablePersonalVaultExportPolicy extends BasePolicy {
 }
 
 @Component({
-  selector: "policy-disable-personal-vault-export",
   templateUrl: "disable-personal-vault-export.component.html",
-  standalone: false,
+  imports: [SharedModule],
 })
 export class DisablePersonalVaultExportPolicyComponent extends BasePolicyComponent {}
