+++ bitwarden_license/bit-web/src/app/admin-console/policies/index.ts
@@ -1,0 +1,1 @@
+export { policyRegister } from "./policy-register";
