+++ bitwarden_license/bit-web/src/app/app.module.ts
@@ -6,20 +6,19 @@ import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
 import { RouterModule } from "@angular/router";
 
 import { JslibModule } from "@bitwarden/angular/jslib.module";
+import { safeProvider } from "@bitwarden/ui-common";
+import { POLICY_REGISTER_TOKEN } from "@bitwarden/web-vault/app/admin-console/organizations/policies";
 import { CoreModule } from "@bitwarden/web-vault/app/core";
 import { OssRoutingModule } from "@bitwarden/web-vault/app/oss-routing.module";
 import { OssModule } from "@bitwarden/web-vault/app/oss.module";
 import { WildcardRoutingModule } from "@bitwarden/web-vault/app/wildcard-routing.module";
 
 import { OrganizationsModule } from "./admin-console/organizations/organizations.module";
-import { ActivateAutofillPolicyComponent } from "./admin-console/policies/activate-autofill.component";
-import { AutomaticAppLoginPolicyComponent } from "./admin-console/policies/automatic-app-login.component";
-import { DisablePersonalVaultExportPolicyComponent } from "./admin-console/policies/disable-personal-vault-export.component";
-import { MaximumVaultTimeoutPolicyComponent } from "./admin-console/policies/maximum-vault-timeout.component";
+import { policyRegister } from "./admin-console/policies";
 import { AppRoutingModule } from "./app-routing.module";
 import { AppComponent } from "./app.component";
-import { FreeFamiliesSponsorshipPolicyComponent } from "./billing/policies/free-families-sponsorship.component";
 import { AccessIntelligenceModule } from "./dirt/access-intelligence/access-intelligence.module";
+
 
 /**
  * This is the AppModule for the commercial version of Bitwarden.
@@ -44,14 +43,13 @@ import { AccessIntelligenceModule } from "./dirt/access-intelligence/access-inte
     RouterModule,
     WildcardRoutingModule, // Needs to be last to catch all non-existing routes
   ],
-  declarations: [
-    AppComponent,
-    DisablePersonalVaultExportPolicyComponent,
-    MaximumVaultTimeoutPolicyComponent,
-    ActivateAutofillPolicyComponent,
-    AutomaticAppLoginPolicyComponent,
-    FreeFamiliesSponsorshipPolicyComponent,
-  ],
+  declarations: [AppComponent],
   bootstrap: [AppComponent],
+  providers: [
+    safeProvider({
+      provide: POLICY_REGISTER_TOKEN,
+      useValue: policyRegister,
+    }),
+  ],
 })
 export class AppModule {}
